refactor: Alert 버튼의 타이틀을 따로 Text 타입으로 구현
