+++ JuiceMaker/JuiceMaker/Controller/JuiceMakerViewController.swift
@@ -89,7 +89,7 @@ class JuiceMakerViewController: UIViewController {
     func currentStockLabelUpdate(fruit: Fruit, label: UILabel) {
         do {
             let stock = try FruitStore.shared.stock(fruit: fruit)
-            label.Message = stock.description
+            label.text = stock.description
         } catch let error as RequestError {
             showNotificationAlert(message: error.errorDescription)
         } catch {
@@ -112,7 +112,7 @@ class JuiceMakerViewController: UIViewController {
         }
     }
     
-    func showNotificationAlert(message: String, title: String = Message.ok.description) {
+    func showNotificationAlert(message: String, title: String = Text.ok.title) {
         let alert = UIAlertController(title: nil, message: message, preferredStyle: .alert)
         let cancel = UIAlertAction(title: title, style: .cancel, handler: nil)
         alert.addAction(cancel)
@@ -121,8 +121,8 @@ class JuiceMakerViewController: UIViewController {
     
     func showOutOfStockAlert() {
         let alert = UIAlertController(title: nil, message: Message.outOfStock.description, preferredStyle: .alert)
-        let cancelAction = UIAlertAction(title: Message.cancel.description, style: .cancel, handler: nil)
-        let okAction = UIAlertAction(title: Message.ok.description, style: .default) { _ in
+        let cancelAction = UIAlertAction(title: Text.cancel.title, style: .cancel, handler: nil)
+        let okAction = UIAlertAction(title: Text.ok.title, style: .default) { _ in
             guard let viewController = self.storyboard?.instantiateViewController(withIdentifier: "FruitStoreView") else { return }
             self.present(viewController, animated: true, completion: nil)
         }
